refactor: move starknet hash function code
