+++ crates/starknet-committer/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod block_committer;
+pub mod hash;
 pub mod starknet_forest;
 pub mod starknet_patricia_merkle_tree;
